refactor: organize content into assets
