+++ .suite-cli/cli/scripts/assets/configIndexContent.asset.js
@@ -1,0 +1,5 @@
+module.exports = ()=>`
+module.exports.config = require('./config')
+module.exports.logger = require('./logger')
+module.exports.morgan = require('./morgan')
+`;
