+++ .suite-cli/cli/scripts/assets/controllersContent.asset.js
@@ -1,0 +1,9 @@
+module.exports = ({ answers }) => `
+const services = require('../services')
+const { asyncErrorHandler, APIError } = require('${answers.project_base}/utilities')
+const hello = asyncErrorHandler(async (req, res) => {
+    res.status(200).json({ data:'Hello from ${answers.project_base}' })
+})
+module.exports = {
+    hello,
+}`;
