+++ .suite-cli/cli/scripts/assets/envContent.asset.js
@@ -1,0 +1,4 @@
+module.exports=()=>`
+PORT=9001
+DATABASE_URL=mongodb://localhost:27017
+`;
