+++ .suite-cli/cli/scripts/assets/errorIndexContent.asset.js
@@ -1,0 +1,3 @@
+module.exports = () => `
+module.exports.errorHandler = require('./errors.handler')
+`;
