+++ .suite-cli/cli/scripts/assets/errorReadmeContent.asset.js
@@ -1,0 +1,7 @@
+module.exports = ({ answers }) => `
+# ${answers.project_base.charAt(0).toUpperCase() + answers.project_base.slice(1)}
+
+## Error handler package
+
+Welcome to the ${answers.project_base.charAt(0).toUpperCase() + answers.project_base.slice(1)} Error Handler Package! This package, residing within our monorepo, provides robust error handling capabilities tailored for microservices architecture. It equips you with the tools to effectively manage and respond to errors across your microservices applications, ensuring reliability and resilience.
+`;
