+++ .suite-cli/cli/scripts/assets/index.js
@@ -1,0 +1,41 @@
+module.exports.ecosystemContent = require('./ecosystemContent.asset')
+module.exports.envContent = require('./envContent.asset') 
+module.exports.apolloServerContent = require('./apolloServerContent.asset') 
+module.exports.k8sClusterDeploymentContent = require('./k8sClusterDeploymentContent.asset') 
+module.exports.k8sIngressServiceContent = require('./k8sIngressServiceContent.asset') 
+module.exports.k8sClusterIpServiceContent = require('./k8sClusterIpServiceContent.asset') 
+module.exports.k8sClientPodContent = require('./k8sClientPodContent.asset') 
+module.exports.k8sClientNodeContent = require('./k8sClientNodeContent.asset') 
+module.exports.k8sReadmeContent = require('./k8sReadmeContent.asset') 
+module.exports.snapshotTestContent = require('./snapshotTestContent.asset') 
+module.exports.integrationTestContent = require('./integrationTestContent.asset') 
+module.exports.unitTestContent = require('./unitTestContent.asset') 
+module.exports.e2eTestContent = require('./e2eTestContent.asset') 
+module.exports.middlewaresIndexContent = require('./middlewaresIndexContent.asset') 
+module.exports.middlewaresReadmeContent = require('./middlewaresReadmeContent.asset') 
+module.exports.utilitiesAsyncErrorHandlerContent = require('./utilitiesAsyncErrorHandlerContent') 
+module.exports.utilitiesReadmeContent = require('./utilitiesReadmeContent.asset') 
+module.exports.utilitiesValidateContent = require('./utilitiesValidateContent.asset') 
+module.exports.utilitiesPickContent = require('./utilitiesPickContent.asset') 
+module.exports.utilitiesIndexContent = require('./utilitiesIndexContent.asset') 
+module.exports.apiErrorContent = require('./apiErrorContent.asset') 
+module.exports.errorIndexContent = require('./errorIndexContent.asset') 
+module.exports.errorHandlerContent = require('./errorHandlerContent.asset') 
+module.exports.genericPackageJsonContent = require('./genericPackageJsonContent.asset') 
+module.exports.configReadmeContent = require('./configReadmeContent.asset') 
+module.exports.configMorganContent = require('./configMorganContent.asset') 
+module.exports.configLoggerContent = require('./configLoggerContent.asset') 
+module.exports.configIndexContent = require('./configIndexContent.asset') 
+module.exports.configConfigContent = require('./configConfigContent.asset') 
+module.exports.controllersIndexContent = require('./controllersIndexContent.asset') 
+module.exports.routesIndexContent = require('./routesIndexContent.asset') 
+module.exports.modelIndexContent = require('./modelIndexContent.asset') 
+module.exports.servicesIndexContent = require('./servicesIndexContent.asset') 
+module.exports.servicesContent = require('./servicesContent.asset') 
+module.exports.routesContent = require('./routesContent.asset') 
+module.exports.controllersContent = require('./controllersContent.asset') 
+module.exports.modelContent = require('./modelContent.asset') 
+module.exports.testContent = require('./testContent.asset') 
+module.exports.serverContent = require('./serverContent.asset') 
+module.exports.rootPackageJsonContent = require('./rootPackageJsonContent.asset') 
+module.exports.errorReadmeContent = require('./errorReadmeContent.asset')
