+++ .suite-cli/cli/scripts/assets/integrationTestContent.asset.js
@@ -1,0 +1,9 @@
+module.exports = () => `
+describe('Integration Tests', () => {
+    test('Example Integration Test', async () => {
+        // Write your integration test logic here
+    });
+
+    // Add more integration tests as needed
+});
+`;
