+++ .suite-cli/cli/scripts/assets/k8sClientNodeContent.asset.js
@@ -1,0 +1,14 @@
+module.exports = () => `
+apiVersion: v1
+kind: Service
+metadata:
+  name: supplier-service-port
+spec:
+  type: NodePort
+  ports:
+    - port: 8008
+      targetPort: 9008
+      NodePort: 31008
+  selector:
+    app: ecommerce-app
+`;
