+++ .suite-cli/cli/scripts/assets/k8sClusterDeploymentContent.asset.js
@@ -1,0 +1,3 @@
+module.exports = () => `
+#TODO: to add Deployment logic here
+`;
