+++ .suite-cli/cli/scripts/assets/k8sClusterIpServiceContent.asset.js
@@ -1,0 +1,3 @@
+module.exports = () => `
+#TODO: to add cluster IP service here
+`;
