+++ .suite-cli/cli/scripts/assets/k8sIngressServiceContent.asset.js
@@ -1,0 +1,3 @@
+module.exports = () => `
+#TODO: to add Ingress service here
+`;
