+++ .suite-cli/cli/scripts/assets/middlewaresIndexContent.asset.js
@@ -1,0 +1,5 @@
+module.exports = () => `
+module.exports.middleware1 = (req,res,next)=>{
+next()
+}
+`;
