+++ .suite-cli/cli/scripts/assets/middlewaresReadmeContent.asset.js
@@ -1,0 +1,7 @@
+module.exports = ({ answers }) => `
+# ${answers.project_base.charAt(0).toUpperCase() + answers.project_base.slice(1)}
+
+## Middlewares Package
+
+Welcome to the ${answers.project_base.charAt(0).toUpperCase() + answers.project_base.slice(1)} Middlewares Package! This package, residing within our monorepo, provides a collection of middleware functions tailored for microservices architecture. These middleware functions enable you to enhance and extend the functionality of your microservices applications with ease.
+`;
