+++ .suite-cli/cli/scripts/assets/modelIndexContent.asset.js
@@ -1,0 +1,1 @@
+module.exports = () => 'const models = require(\'./model\');\nmodule.exports = models;';
