+++ .suite-cli/cli/scripts/assets/routesContent.asset.js
@@ -1,0 +1,1 @@
+module.exports = () => 'const express = require(\'express\');\nconst router = express.Router();\nmodule.exports = { router };';
