+++ .suite-cli/cli/scripts/assets/routesIndexContent.asset.js
@@ -1,0 +1,1 @@
+module.exports = () => 'const { router } = require(\'./routes\');\nmodule.exports = { router };';
