+++ .suite-cli/cli/scripts/assets/servicesContent.asset.js
@@ -1,0 +1,1 @@
+module.exports = () => 'module.exports = {};';
