+++ .suite-cli/cli/scripts/assets/servicesIndexContent.asset.js
@@ -1,0 +1,1 @@
+module.exports = () => 'module.exports = {};';
