+++ .suite-cli/cli/scripts/assets/unitTestContent.asset.js
@@ -1,0 +1,9 @@
+module.exports = () => `
+describe('Unit Tests', () => {
+    test('Example Unit Test', () => {
+        // Write your unit test logic here
+    });
+
+    // Add more unit tests as needed
+});
+`;
