+++ .suite-cli/cli/scripts/assets/utilitiesAsyncErrorHandlerContent.js
@@ -1,0 +1,9 @@
+module.exports = () => `
+const asyncErrorHandler = (fn) => (req, res, next) => {
+    return Promise.resolve(fn(req, res, next)).catch((err) => {
+        next(err)
+    })
+}
+
+module.exports = asyncErrorHandler
+`;
