+++ .suite-cli/cli/scripts/assets/utilitiesIndexContent.asset.js
@@ -1,0 +1,6 @@
+module.exports = () => `
+module.exports.APIError = require('./APIError')
+module.exports.asyncErrorHandler = require('./asyncErrorHandler')
+module.exports.pick = require('./pick')
+module.exports.validate = require('./validate')
+`;
