+++ .suite-cli/cli/scripts/assets/utilitiesReadmeContent.asset.js
@@ -1,0 +1,8 @@
+module.exports = ({ answers }) => `
+# ${answers.project_base.charAt(0).toUpperCase() + answers.project_base.slice(1)}
+
+## Utilities
+
+Welcome to the ${answers.project_base.charAt(0).toUpperCase() + answers.project_base.slice(1)} Utilities Package! This package is a part of the larger monorepo dedicated to housing reusable utility functions tailored for microservices architecture. Within this package, you'll find a collection of utility functions designed to streamline various aspects of microservices development.
+
+`;
